strangler fig done

+++ user-service/src/main/java/es/codeurjc/users/service/UserService.java
@@ -1,7 +1,11 @@
 package es.codeurjc.users.service;
 
+import es.codeurjc.users.client.CommentClient;
+import es.codeurjc.users.exception.UserHasCommentsException;
 import es.codeurjc.users.model.User;
 import es.codeurjc.users.repository.UserRepository;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -10,10 +14,13 @@ import java.util.Optional;
 @Service
 public class UserService {
 
+	Logger logger = LoggerFactory.getLogger(UserService.class);
 	private final UserRepository users;
+	private final CommentClient commentClient;
 
-	public UserService(UserRepository users) {
+	public UserService(UserRepository users, CommentClient commentClient) {
 		this.users = users;
+		this.commentClient = commentClient;
 	}
 
 	public void save(User user) {
@@ -37,9 +44,15 @@ public class UserService {
 		return users.existsById(id);
 	}
 
-	public void deleteById(long id) {
-		// check in monolith amount of comments
-		users.deleteById(id);
+	public User deleteById(long id) {
+		User user = users.findById(id).orElseThrow();
+		if (!commentClient.userHasComments(id)) {
+			logger.info("user has 0 comment, then it will be deleted");
+			users.deleteById(id);
+			return user;
+		}
+		logger.info("user has comments, therefore it will not be deleted");
+		throw new UserHasCommentsException();
 	}
 
 	public Optional<User> findByNick(String nick) {
